Split vpPylon* classes info multiple files
One class per file.
